samples/python/agents/a2a_end2end_oauth/pyproject.toml
[project]
name = "airbub-ma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.8",
    "click>=8.2.1",
    "dotenv>=0.9.9",
    "geopy>=2.4.1",
    "google-adk>=1.3.0",
    "google-auth-oauthlib>=1.2.2",
    "gradio>=5.34.0",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.25",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "uvicorn>=0.34.3",
]
